Better code splitting in util script
